+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ** (master) **
 
+- Add Sé shape file
+
 **1.0.1**
 
 - Modularize mode choice via injection and configuration